+++ btree/btree_config.h
@@ -1,0 +1,14 @@
+#ifndef __BTREE_CONFIG_H__
+#define __BTREE_CONFIG_H__
+
+/*
+ * defines the minimum filled elements in one node
+ */
+#define ORDER 2
+
+/*
+ * defines the maximum filled elements in one node
+ */
+#define MAXNODE (2 * ORDER)
+
+#endif //__BTREE_CONFIG_H__
