--- insert.h
@@ -1,17 +1,0 @@
-/*
- * insert.h
- *
- *  Created on: Nov 11, 2016
- *      Author: Matthias Nickel
- */
-
-#ifndef INSERT_H_
-#define INSERT_H_
-
-#include "btree.h"
-
-short insert(stdelement e);
-void insert_into_node(struct node* node, stdelement e);
-short getIndex(struct node* node, stdelement e);
-
-#endif /* INSERT_H_ */
